Move common types and schema to case folder
