+++ x-pack/plugins/actions/server/builtin_action_types/case/common_schema.ts
