+++ x-pack/plugins/actions/server/builtin_action_types/jira/mocks.ts
@@ -6,7 +6,7 @@
 
 import { ExternalService, PushToServiceApiParams, ExecutorSubActionPushParams } from './types';
 
-import { MapRecord } from '../case/types';
+import { MapRecord } from '../case/common_types';
 
 const createMock = (): jest.Mocked<ExternalService> => {
   const service = {
